+++ src/sargeom/coordinates/ellipsoids.py
@@ -35,7 +35,7 @@ class Ellipsoid:
         self._n = self._f / (2 - self._f)  # Third flattening factor
         self._e2 = self._f * (2 - self._f)  # Eccentricity squared
         self._e = np.sqrt(self._e2)  # Eccentricity
-        
+
         n = [self._n**i for i in range(11)] # n^0 ... n^10
         # Coefficients d(n) of the series expansion of the inverse conformal latitude
         self._phichi = (
@@ -90,12 +90,12 @@ class Ellipsoid:
         ----------
         phi : array_like
             The geodetic latitude in radians.
-        
+
         Returns
         -------
         nu : :class:`numpy.ndarray`
             Prime vertical curvature radius in meters.
-            
+
         Notes
         -----
         The prime vertical curvature radius is the radius of curvature in the 
@@ -124,7 +124,7 @@ class Ellipsoid:
         ----------
         phi : array_like
             The geodetic latitude in radians.
-            
+ 
         Returns
         -------
         psi : :class:`numpy.ndarray`
@@ -159,7 +159,7 @@ class Ellipsoid:
         ----------
         psi : array_like
             The isometric latitude in radians.
-            
+
         Returns
         -------
         phi : :class:`numpy.ndarray`
@@ -231,7 +231,7 @@ class Ellipsoid:
         ----------
         chi : array_like
             The conformal latitude in radians.
-            
+
         Returns
         -------
         phi : :class:`numpy.ndarray`
@@ -315,7 +315,7 @@ class Ellipsoid:
             The Y cartesian ECEF coordinate in meters.
         Z : array_like
             The Z cartesian ECEF coordinate in meters.
-        
+
         Returns
         -------
         lamb : :class:`numpy.ndarray`